+++ scripts/installers/zellij.sh
@@ -11,17 +11,17 @@ BIN_PATH="zellij"
 INSTALL_NAME="zellij"
 
 # Make sure that ~/.local/bin exists
-mkdir -p ${DOWNLOAD_DEST}
-mkdir -p ${INSTALL_DEST}
+mkdir -p "${DOWNLOAD_DEST}"
+mkdir -p "${INSTALL_DEST}"
 
-# Install pre-built binary of nvim
-curl -LO https://github.com/zellij-org/zellij/releases/latest/download/${TARBALL_NAME}${TARBALL_EXT}
+# Install pre-built binary of zellij
+curl -LO "https://github.com/zellij-org/zellij/releases/latest/download/${TARBALL_NAME}${TARBALL_EXT}"
 
 # Extract and untar, move to the specified download folder, and remove the download
-tar ${TAR_ARGS} ${TARBALL_NAME}${TARBALL_EXT} 1>/dev/null
-rm -rf ${DOWNLOAD_DEST}/${TARBALL_NAME} 2>/dev/null
-mv -f ${BIN_PATH} ${INSTALL_DEST}/${INSTALL_NAME}
-rm -f ${TARBALL_NAME}${TARBALL_EXT}
+tar "${TAR_ARGS}" "${TARBALL_NAME}${TARBALL_EXT}" 1>/dev/null
+rm -rf "${DOWNLOAD_DEST:?}/${TARBALL_NAME}" 2>/dev/null
+mv -f "${BIN_PATH}" "${INSTALL_DEST}/${INSTALL_NAME}"
+rm -f "${TARBALL_NAME}${TARBALL_EXT}"
 
 # Give it executable permission
-chmod +x ${INSTALL_DEST}/${INSTALL_NAME}
+chmod +x "${INSTALL_DEST}/${INSTALL_NAME}"
